+++ Editor/Processors/SkinnedMeshes/RemoveMeshInBoxProcessor.cs
@@ -1,5 +1,7 @@
+using System;
 using System.Collections.Generic;
 using System.Linq;
+using Anatawa12.AvatarOptimizer.ErrorReporting;
 using nadena.dev.ndmf;
 using UnityEngine;
 
@@ -24,27 +26,9 @@ namespace Anatawa12.AvatarOptimizer.Processors.SkinnedMeshes
                     inBoxVertices.Add(vertex);
             }
 
+            Func<Vertex[], bool> condition = primitive => primitive.All(inBoxVertices.Contains);
             foreach (var subMesh in target.SubMeshes)
-            {
-                int srcI = 0, dstI = 0;
-                for (; srcI < subMesh.Triangles.Count; srcI += 3)
-                {
-                    // process 3 vertex in sub mesh at once to process one polygon
-                    var v0 = subMesh.Triangles[srcI + 0];
-                    var v1 = subMesh.Triangles[srcI + 1];
-                    var v2 = subMesh.Triangles[srcI + 2];
-
-                    if (inBoxVertices.Contains(v0) && inBoxVertices.Contains(v1) && inBoxVertices.Contains(v2))
-                        continue;
-
-                    // some vertex is not in box: 
-                    subMesh.Triangles[dstI + 0] = v0;
-                    subMesh.Triangles[dstI + 1] = v1;
-                    subMesh.Triangles[dstI + 2] = v2;
-                    dstI += 3;
-                }
-                subMesh.Triangles.RemoveRange(dstI, subMesh.Triangles.Count - dstI);
-            }
+                subMesh.RemovePrimitives("RemoveMeshInBox", condition);
 
             // We don't need to reset AdditionalTemporal because if out of box, it always be used.
             // Vertex.AdditionalTemporal: 0 if unused, 1 if used
@@ -52,7 +36,7 @@ namespace Anatawa12.AvatarOptimizer.Processors.SkinnedMeshes
             inBoxVertices.Clear();
             var usingVertices = inBoxVertices;
             foreach (var subMesh in target.SubMeshes)
-            foreach (var vertex in subMesh.Triangles)
+            foreach (var vertex in subMesh.Vertices)
                 usingVertices.Add(vertex);
 
             // remove unused vertices